HUDSON-6814: Advanced approve commands does not get used on save.
The issue appeared because of the testability changes
in GerritNotifier, it got the first instance of the Config object.
Removed all other ways of getting the config besides from PluginImpl,
and changed so that a new Notifier is created with a fresh Config
every time.
There might have sneaked in some synchronization changes in
the BuildMemory, but they have nothing to do with this bug,
only insurance.