Step one for the retrigger feature.
The Build-Cause now lists the other "non silent" builds that was
involved in the event.
Serialization issue has been taken care of.
I did a semi bad thing for this commit
and enabled checkstyle on test sources that made me go and fix
alot of checkstyle warnings in places not belonging to this change.
Updated the checkstyle rules from source.