HUDSON-6818 Retrigger jobs step one.
The user can now retrigger one job for a Gerrit event.
If all jobs are completed, the result from the other jobs
will be used when calculating the end result.
If some jobs are still running for the event, the former job
for the project of the retriggered job will be replaced.
Still TODO: Retrigger All jobs for an event.