Fixfor backwards compatibility issue with TriggeredItemEntity
When the manual trigger feature was implemented a refractor of
TriggeredItemEntity was done, so it could be used from other classes.
That caused CNFs to be thrown during loading of old builds.
This commit fixes that by adding a special converter for the
TriggerContext class.
Also bumped the version numbers since the new version is mentioned in javadoc.