Fixed NPE when 0 projects configured in GerritTrigger.
If a project was loaded/configured with 0 projects,
then the list of projects in GerritTrigger is set to null,
and that would cause a NPE to be thrown in isInteresting.
And all the following triggers would not be called.
This patch fixes that by checking for null in isInteresting
plus a catch around each call to a trigger's isInteresting
to shield from future runtime errors.