HUDSON-7207 Teach the trigger to understand merge commits.
There was a problem getting merge commits to compile when using this plugin,
caused by GerritTriggerBuildChooser.getFirstParent. The reason is that when it
invokes "git log" on merge commits, it will get back more than one SHA1. The
string it gets has the format:

<PARENT1_SHA1> <PARENT2_SHA1>[ <PARENTX_SHA1...]

This patch quite simply splits the returned value by SPACE, and uses
PARENT1_SHA1 to get the commit object.