Fix Ant pattern matching on Windows
When Ant pattern matching is used the Ant SelectorUtils class is used to
determine if there is a match, and the behavior of SelectorUtils depends on
the value of java.io.File.separator.

Replacing '/' with java.io.File.separatorChar before calling SelectorUtils
is necessary to ensure that path matching works the same on all platforms.
Patterns that use the Windows separator ('\') will still match differently
on Windows and Linux, but it is not clear what behavior the user would
expect in that case anyway.