Cancel jobs when a new patch set is created
When a new patch set is posted, building the previous patch set
is no longer valuable. This causes an enormous extra load on the
build servers. If someone rapidly posts several revisions of a
patch, it could cause enormous delays and wasted resources in
Jenkins.

This only does part of the work, because calling cancel() on a
Future that is provided by scheduleBuild2() does not cancel jobs
that are in the queue. A ticket has been created to address this,
available at: https://issues.jenkins-ci.org/browse/JENKINS-12152