[JENKINS-11726] Do not count NOT_BUILT as a fail.
With this change, NOT_BUILT changes are ignored when tallying up what code
review / verified score to give a change. If none of the builds complete
(they're all "NOT_BUILT"), then we use the score & command assigned to the
NotBuilt event.

This is required for plugins to be able to "skip" some gerrit verifiers. (like
the pathignore-plugin)