Synchronization and NPE fix of runningJobs.
Null checks and synchronization of the map of running jobs.
There where problems when the trigger was deserialized,
that could cause NPE to be thrown.
Also added synchronization to avoid dirty writes.

I still get unexpected behaviour with this feature;
builds fail to cancel and it somethimes ends up
in a strage state, so I put it as more optional
than before and changed the default to be off.

Change-Id: Ic3f135f2c961161d35c2067d6c0769e7167668b2