JENKINS-12881: Obtain failure message at completed-time.
Resolves a potential race condition when a single job
may be executing multiple times before the change "batch"
status is ready to be reported back to Gerrit.

The test setup is:

 * Job-A
   - Shell: echo "Starting ${BUILD_TAG}" > error-message.txt
            sleep 5
   - Shell: echo "Always fails ${BUILD_TAG}" > error-message.txt
            sleep 5
            exit 1

 * Job-B
   - Shell: echo "Starting ${BUILD_TAG}" > error-message.txt
            sleep 15
   - Shell: echo "Never fails ${BUILD_TAG}" > error-message.txt
            sleep 15

Job-B will always be the long-pole (30 sec). Job-A will always fail,
hopefully with "Always fails ______" as the error-message.

Create 2 commits for review, and push both commits to Gerrit. Both
Job-A builds will run and fail within 20 seconds, before the first
Job-B build is even finished executing.

The test is successful if both review commits contain the correct
"Always fails {TAG}" messages, corresponding to the correct URL.
The test fails if any other string exists in the failure message.