Fixed a double-trigger when a project is renamed.
GerritHandler.gerritEventListeners uses a hash code as the key to
remember listeners, and GerritTrigger derives hash code from the project
name.

So when a project is renamed, the hash code changes, and
GerritTrigger.stop() fails to remove the listener.

This test case and the fix resolves this issue with the minimal change
to the current code. It uses the project object itself as the source of
the hash, which works correctly.

This change by itself doesn't touch the bigger underlying design
problem, that the hash code is used as a key. If two projects have the
exact same hash code for coincidence (which is legal), then the current
code doesn't work.

I'm providing this change as the first step to maximize the chance of
the change inclusion.