Workaround for parameter issues on Windows
I've found that in some configurations (particularly with Gradle and on
Windows) that there are problems with the way some parameters are passed
on the commandline (to Gradle). I've not been able to sort out the exact
cause of the problem (I originally attempted to fix by changing the way
things were escaped but was unsuccessful). Essentially the problem I've
seen is that gradle can't process the commandline arguements correctly
when triggered by this plugin (but not otherwise and I don't see the
same problems with, say, the Maven plugin).

To work around the problem I've added another checkbox under the
advanced options which stops this plugin from setting the 'name and
email' parameters for a build. Without those parameters things work
nicely for me. I'd rather not maintain a fork just for my company so I'd
like to get this workaround (or something like it) merged if possible
(let me know if there is anything I can do help out with that).

I think this also fixes an issue where it appeared that manually triggered
changes were not respecting the escape quotes setting (it seemed to be
forced to false but it's also entirely possible that I was missing
another source the setting).