Nullpointer when project cant be resloved
When getting Gerrit parameters the project is needed to sort
out if Name/Email should be fetched or if escapeQoutas is needed.

But if the project doesn't exist when triggered manually a
nullpointer is thrown.

Added nullpointer check and default values if project is missing.
Added method setOrCreateParametes without parameter project

Change-Id: I72fc2812073772024d837d392b09e34bf041aac9