Rename constants for better readability and consistency
The constants `PROTO` and `GERRIT_PROTO` are renamed to `SCHEME` and
`GERRIT_SCHEME`.  This makes the naming of all the provider constants
consistent with RFC3986 [1].

The `proto` field is replaced with `scheme`.  The methods to get and
set `proto` are deprecated and new methods to get/set `scheme` are
added.

JSON parsing is modified to use `proto` if it exists, or otherwise
use `scheme`.

[1] http://tools.ietf.org/html/rfc3986

Change-Id: I38890355daa5d614e6b43fd351b199ed4f9e83c1