Copy setting in PluginConfig to Config
NumberOfSendingWorkerThreads is moved to PluginConfig. But ones in Config
is still needed for several process. Unfortunately it is difficult to
replace easily.

So copy value from PluginConfig to Config whenever Config instance is
created and PluginConfig instance is updated.

Note that this is workaround. We should adopt another solution in future.