Refactored REST api code to gerrit-events
Most of the code should belong there.
In order to make this work a bit more smoothly some of
the configuration interfaces were also refactored to
better split the difference between the two ways, but
still have common code.
As the interfaces might break backwards compatibility
with other tools using stream-events the version was bumped
all the way up to 1.99 to indicate that it is just a hair
away from 2.0, where the module should be broken free.

Change-Id: Ie1662a3354140461af2aa098d796b27eb9f093f2