Add data structure and UI Support for replication.
The admin can enable replication in server configurations,
add Gerrit slaves, and enable slave selection in jobs.
If the latter is enabled, the user will see a dropdown
containing a list of Gerrit slaves configured with the
Gerrit server that has been selected, and can choose one
slave after which the job will wait for replication events.

Change-Id: Id08852619442290b3b5c8e2f179d10989e98b1af