Add option to exclude drafts when triggering on patchset created event.
Gerrit support draft changes and draft patchsets. Gerrit Trigger was able
to trigger a build when a draft change is published (transitioned from draft to
regular change) but it was missing an option to allow user to exclude draft
patchsets when triggering on patchset created event.

This change add a checkbox in the patchset created type of trigger to exclude
draft patchsets. By default, this option is not enabled to be compatible with behaviour
prior to this change.

Change-Id: I279a34662a46c6aa9657b2b604027b0e52f76adc