Include selected labels to review input for REST API
Message for REST API always includes labels for both code review and
verified. But current Gerrit does not have ones for verified by default.
It causes that review would be denied if label for verified is not added
to gerrit.

This patch can add user selected labels to message for REST API.

Fix for JENKINS-24074

Task-Url: https://issues.jenkins-ci.org/browse/JENKINS-24074