Add one-off executor to search list for cancel job
By "Build Current Pathcset Only" feature, running and queued jobs with
the same change are canceled. These are found in queue items and
executors.

But MatrixJob uses one-off executor. It is not included in executors. So
it cannot be canceled.

This patch adds one-off executors to executor's list.

Fix for JENKINS-24295

Task-Url: https://issues.jenkins-ci.org/browse/JENKINS-24295