rescheduled inflight pushes + events not in cache
With high volume repos, it is possible for Gerrit to re-schedule a replication
due to an inflight push. This re-schedule will result in a failed replication state
and will result in a failed build. By only proceeding on a successful replication
status, this will be avoided.

This commit also includes a fix to cope with missed events due to a Jenkins restart