Correct infinite loops caused by CommentAdded
CommentAdded events were not be processed with the correct
gerritEvent() signature. This caused builds be triggered needlessly
and sometimes caused infinite loops between Jenkins and Gerrit.

This fix exposes the EventListener as a public class so that the Gerrit Events
Gerrit Handler can correctly invoke the appropriate methods.

This also corrects a NPE during a BuildStarted notification.

[JENKINS-23152]

Change-Id: Ief95fc05d5cf0a5d2e0dc95330e74bb6d52f425c