Introduce Google Truth
Truth [1] is a testing framework designed to make tests and their error
messages more readable and discoverable, while being extensible to new
types of objects.

Add dependency on truth in the pom file and convert some of the trigger
tests to use it.

This also gets rid of several deprecation warnings which were caused
by using junit.framework.Assert

[1] http://google.github.io/truth/

Change-Id: I6e2ed8e646fa4831a41639ad8a12b5b0a3d98f03