Build with Java 7
Public updates for Java 6 ended in Feb 2013 [1].

Update the project to build with Java 7.

Java 7 has support for type inference, so we no longer need to give
explicit specification of type arguments.  For example code like:

   List<String> list = new List<String>();

will now cause a warning "Redundant specification of type arguments".

Code using this syntax can be changed to:

   List<String> list = new List<>();

which silences the warning and makes the code a bit more compact.

[1] http://www.oracle.com/technetwork/java/eol-135779.html
[2] http://docs.oracle.com/javase/tutorial/java/generics/genTypeInference.html

Change-Id: Ifd72081ae7d6090594cd3d5cb219a639da64424a