Change should be checked for BuildCurrentPatchsets
If Build Current Patchsets only is enabled, then the Change in question must first be
checked to see if a build should be aborted.

Without this fix, it is possible for a build to be aborted because of a NON-related patchset
from ANOTHER change.

This change also makes the two new options disabled by default to be compatible
with the previous behaviour.

Change-Id: I958a4974756f398d87c7d60fb23cdd1b978a0ea1