add actions to schedule2 for Workflow Jobs
Add the badgeAction, RetriggerActions, and paramter actions for the
Workflow Job. It may be worth just using straight up SCMTrigger
interface for both workflow and abstract projects, as this interface is
supported by both, however this patch does not make this change, as we
lose the ability to pause the Cause directly into the AbstractProject's
ScheduleBuild2 function.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>