Option to permit disable Strict Forbidden files
Add option to gerrit project to permit the disable of the default
strict forbidden file verification of an event

Enabling this option will allow an event to trigger a build if the event contains BOTH one or more wanted file paths/topics/branches AND one or more forbidden file paths.

In other words, with this option, the build will not get triggered if the change contains only forbidden files, otherwise it will get triggered.

[FIXED JENKINS-30620]

Change-Id: Icb84111c82a38db722ec09bf7538441bc1b9424a