Update for change to Gerrit stream events
A pending Gerrit stream events change[1] will append approval info
to every comment-added event.  That change also adds an `updated`
attribute to the approvals info.  This change will ensure that the label
trigger will continue to work with existing Gerrit versions and with the
upcoming stream events change in newer Gerrit versions.

This change dependings on a pending gerrit-events update[2]

[1] https://gerrit-review.googlesource.com/#/c/71051
[2] https://github.com/sonyxperiadev/gerrit-events/pull/48