Add Replication status to queued item and failures
This change adds the replication status to the reason for the fact
that an item is in the queue.

For example, if a replication is failig to a mirror, it will added as show below:

(pending—Replication to sekimirror (is failing) not completed.)

In addition, if the replication has timed-out, the status will be displayed in the console log
as shown below:

ERROR: Waiting for replication of refs/changes/55/899655/1 to sekimirror timed out.

Replication Status:
-------------------

** Replication to sekimirror is failing!

Change-Id: I254d006629ddae375a3b95ad9a1b137c4a8006be