Correct Missed Events Playback manager init
It is possible for the Missed Events playback manager to not be correctly initialized
once the REST API has been enabled AFTER the connection has been started.

The same issue applies if the Gerrit events-log plugin was installed AFTER the
Gerrit connection was started.

This ensures that the listener is active so that playback can execute when required.

[FIXED JENKINS-31439]