Only print the parameters in the test ParameterBuilder
Since jenkins test harness is populating the builds with the environment
of the process the pull request subject was also added to the unit test build.
That caused one of the tests to fail because it was looking for the absence of a
parameter name that was present in the pull request title.
Changing the test builder to only print the parameters should solve it.