More test tweaking
Some tests had lost their @Test annotation when being migrated
to JenkinsRule. Enabled and fixed them.
Fixed SshdServerMock to be able to run in Parallel and fixed
tests to adapt to it.
Migrated GerritTriggerTimer to use jenkins.util.Timer instead of
java.util.Timer because it was breaking some tests.
Tweaked sure fire to not reuse forks and the max heap for each suite.