Provide GERRIT_MERGED_REVISION after change-merged
The newRev field of a change-merged event was not parsed yet.
Since this field holds the hash of a merge commit it can be important
in certain build conditions.
This patch exposes this field as a GERRIT_MERGED_REVISION variable
to Jenkins builds. It requires another patch in gerrit-events to work.