[JENKINS-30821] Added comment as job parameter
Expose the comment posted to Gerrit on a comment-added event as a job
parameter named GERRIT_CHANGE_COMMENT_TEXT.  Use the option for commit
message encoding also for this parameter for now.  This allows using the
comment for logic in the job script (in particular, for pipeline jobs
this enables all kinds of fancy setups).