Explicitly provide GerritTrigger object to `schedule` method
Before it was possible to hit the race condtion during Jenkins
initalization: GerritTrigger was already completely initialized, but Job
object was still "under construction". As result
`calculateDynamicConfigRefreshInterval` did a fallback to Average
interval. Unfortnately, the list of server could be non-initilized as
well. That leads to using `MINIMUM_DYNAMIC_CONFIG_REFRESH_INTERVAL`
instead of specified one.