Added locking around `dynamicGerritProjects`.
In my Jenkins instance, I seemingly randomly have these instances where two
patchset-created events, sent seconds apart, have a problem: one of them
will be processed normally, and the other will not.

After a whole bunch of logging and debugging, I have traced the problem to
the `dynamicGerritProjects` list.  On the one that works, the list will be
the proper size; however, on the one that doesn't, it'll be empty.

I suspect that this has to do with the task that updates the dynamic trigger
configuration in the background at the moment when the new event comes in.
Because the timing code is all coordinated by Jenkins, it is at least
plausible that race conditions could be more likely than expected, since
tasks will tend to run at whatever internal intervals Jenkins uses for them.

The logging changes generally add some context to otherwise context-free
log statements.

The locking around `dynamicGerritProjects` aims to control every access of
that variable.  For some strange reason, making the lock a normal class member
did not work (it was always null, even when intantiated in the definition as
well as the two constructors).  However, using a static constant object seems
to have at least stopped all the null-pointer exceptions.

I'm currently testing this on my instance, and we tend to have this problem
come up once or twice a week.  We'll see how it goes.