Trigger restarts should not have a bad initial state.
The `start` action here doesn't actually fetch the dynamic project configuration.
Instead, it starts a thread to do it.  This means that there is a window of time
between when `start` returns and when the thread completes its first task that the
dynamic project list is empty.

For classic use, this was probably not a huge issue.  However, using imperative
Jenkinsfile jobs, this is a major problem because after every job is run, the
configuration "changes" (if you use the `properties` command in the job).

In our case, when multiple Gerrit events were submitted in rapid succession, some
would be "lost" due to the trigger-restart time window where there were no dynamic
projects listed.

This commit ensures that the dynamic project list is updated as part of the `start`
action so that the trigger is always valid.