Add option to include matrix results in message
Add an option to matrix jobs that will include the build results of all
child builds in the final build result message that is sent to Gerrit.
The option can be found in Matrix Job -> Gerrit Trigger -> Advanced ->
Include Matrix Results. This option will probably produce too much
output with large matrices, but can be very useful with smaller ones.