Do not use synchronised checks during event processing
Using synchronised checks during event processing leads to contention on
GerritWorker threads. Especially if lock is done on ToGerritRunListener.
So, to speed up the processing we need to get rid of
`isProjectTriggeredAndIncomplete`.

`isProjectTriggeredAndIncomplete` can be replaced by `isTriggered` and `isBuilding`.
That transforms the previos version of check to `(isTriggered && isBuilding) || isTriggered`.
So, it's absolutely safe to keep `isTriggered` only.

After that plugin itself does not use `isProjectTriggeredAndIncomplete`.