Use a latch to force EventListener to wait until the project list is ready
The EventListener always "waits" for the project list to be ready before
taking action on an event.  For static configurations, this is always
a no-op, since the latch counter will always be zero.  For dynamic
configurations, the latch will be initialized to "1" when the trigger
is started; as soon as the configuration has been loaded, it will
decrease to zero.

When the job is updated or repeatedly stopped and started, the latch
will only be reset to "1" if the dynamic configuration URL changes.
This way, things can move smoothly if you're using a pipeline job
that calls "properties()" to reset itself every time.