Initialize the latch in `readResolve`
`readResolve` is called when the plugin is loaded from disk, and
this bypasses the constructors (since we don't have a zero-argument
version).  In that case, we need to set up the latch appropriately.

I took it out of the constructors because it was redundant, since
it's defined that way at the class level.

I also added some logging for when the different constructor options
are called.