Introduce custom handling of aborted jobs
Prior to this commit the builds that where aborted would undergo the
same handling as jobs that failed. This is not always the most desirable
behaviour. For instance the gerrit-trigger plugin allows to abort
running jobs when a new patchset is created while building. In such case
the build is aborted and a new build is started. Prior to this commit
this could result in a patchset getting a -1 verified vote causing some
confusion for the user.

If upgrading from a previous version the default behaviour will remain
backwards compatible. To enable custom handing one needs to edit the
configuration and configure the build aborted default values to the
desired value(0).

Also see https://issues.jenkins-ci.org/browse/JENKINS-14767

Original-author: Réda Housni Alaoui
Signed-off-by: Kees Jongenburger <kees.jongenburger@exset.com>