Always unblock executors waiting for dynamic trigger configuration
Executors are waiting for the project to be ready. This is done so the
list of dynamic trigger projects has finished loading. To ensure the
tasks are not waiting forever, the list must always be set and the latch
released, even when the job is disabled or the trigger was disabled,
fetching of the trigger list failed or similar.

Added two new tests, one for failing to fetch the trigger file, one for
disabled job.

Signed-off-by: Georg Bremer <georg.bremer@here.com>