Add listener for cleaning up removed triggers
If a gerrit trigger is removed from a job, then the listener won't be
removed. This caused loads of log entires to be written to the Jenkins
system log, warning that the trigger couldn't be found. E.g.:

> Couldn't find a configured trigger for Some/Job

This change fixes this issue by listening for save events on jobs and
removing listeners if there is no gerrit trigger configured on the
newly updated job.