Fix excessive triggers for a change with forbidden files only
With strictForbiddenFileVerification enabled, when the list of files in
a change was fully covered by forbidden file paths, the change was still
considered interesting. That's because the list of files always contains
at least one magical file name: "/COMMIT_MSG" [1]. In some cases a change
may also contain some other magical file names: "/MERGE_LIST" and/or
"/PATCHSET_LEVEL" (at least as of Gerrit 3.4). So after all, when the
list of files was checked against forbidden file paths, the resuling
list of files (tmpFiles) was never empty and the associated job was
being triggered, although it shouldn't have been.

[1] https://gerrit.googlesource.com/gerrit/+/refs/tags/v3.4.0/java/com/google/gerrit/entities/Patch.java#33