JENKINS-69050: Stop old server instances correctly during JCasC update
When applying a configuration update to the Gerrit Trigger plugin
via "Jenkins Configuration as Code", a reference to the list of
server instances was saved instead of creating a new list. This
meant that after the configuration was updated, the list reference
pointed at the new list of server instances and not the old one,
and hence failed to shutdown the old server connections.

This patch additionally avoids restarting the existing server
connections unnecessarily when "configure()" is called with
"dryrun" set to true.