Increase the default events log support check period
The previous default events log checking period was two seconds. This
spams Gerrit server logs with unnecessary requests checking for this
functionality. Gerrit plugins don't tend to be added and removed quickly
enough where a 2 second interval to check if it is supported is
appropriate. Instead increase the interval to once per day by default.

This is quickly enough that if a Gerrit server is updated you'll notice
in a reasonable amount of time, but infrequent enough to not flood
server logs with annoying 404 request logs for /plugins/events-log/.
Additionally it appears this Jenkins plugin queries for this support on
startup which means Jenkins users can restart the Jenkins installation
to have it query more quickly.