Provide commit message in manual queries [NO-NOT-MERGE]
Enabling getting comments… for testing purposes.

Gerrit 3.4.6 defaults to not show commit messages unless
`--commit-message` is provided. This yields a mismatch between
"automatically" and "manually" triggered changesets in the plugin
(GERRIT_CHANGE_COMMIT_MESSAGE missing in the latter case). This commit
switches to a different overload of GerritQueryHandler's queryJava()
providing this functionality.