JENKINS-72649: Add project parameter to SSH commands
Add --project to the SSH review command calls so that it is possible
to distinguish between conflicting change numbers within distinct projects.
This change is backwards compatible to earlier versions of Gerrit, but
becomes relevant both from optimization reasons since 2.15 and to
prevent change collisions from 3.5 with the support for imported
projects.

Bug: Issue JENKINS-72649
Release-Notes: Add project parameter to SSH commands